add fix for curried theory operators
